Example.java
A text file that contains one or more class definitions. The Java compiler requires that one source file uses the .java filename extension. The name of the class defined by the program is Example. In Java, all the code must reside inside a class and by convention, the name of the main class should match the name of the file that holds the program.