Compiling the Program
To compile the Example program, execute the compiler, javac, specifying, the name of the source file on the command line. The javac compiler creates a file called Example.class that contains the bytecode version of the program. The output of javac is not code that can be directly executed. To actually run the program, you must use the Java application launcher called java. To do so, pass the class name Example as a command-Line argument. C:\>java Example .