The if Statement
The Java if statement works much like the IF statement in any other language. Further, it is syntactically identical to the if statements in C, C++, and C#. Its simplest form is if(condition) statement; . Her, condition is a boolean expression. If condition is true, then the other statement is executed . If condition is false, then the statement is bypassed.