The for Loop
The simplest form of the for loop is for(initialization; condition; iteration) statement; . In its most common form, the initialization portion of the loop sets a loop control variable to an initial value. the condition is a Boolean expression that tests the loop control variable. If the outcome of that test is true, for the loop continues to iterate. If it is false the loop terminates . the iteration expression determines how the loop control variable is changed each time the loop iterates. Her is a short program that illustrates the for loop.