Using Blocks of Code
The following program uses a block of code as the target of a for loop. Java allows two or more statements to be grouped into blocks of code, also called code blocks. This is done by enclosing the statements between opening and closing curly braces. Once a block of code has been created, it becomes a logical unit that can be used any place that a single statement can. for example, a block can be a target for java's if and for statements.