Characters
In Java, the data type used to store characters is char. Java uses Unicode to represent characters. Unicode is a fully international character set that can represent all of the characters found in all human languages. It is a unification of dozens of character sets, such as Latin, greek, arabic, Cyrillic, Hebrew, Katakana, Hangul, and many more.