Booleans
Java has a primitive type, called boolean, for logical values. It can have only one of two possible values, true or false. This is the type returned by all relational operators, as in the case of a < b boolean is also the type required by the conditional expression that govern the control statements such as if and for.