Demonstrate block scope
The scope defined by a method begins with its opening curly brace. However if that method has parameters, they too are included within the method's scope. As a general rule, Variables declared inside a scope are not visible (that is accessible to the code that is defined outside the scope).